Moved place files to a subdirectory
