+++ project_tracking/sprint_backlog.txt
@@ -2,7 +2,7 @@ Third Sprint
 Thurs Sat Jan 26 -- Sat Feb 2
 "Get out'a port"
 
-1/2. Move place stuff to it's own directory
+--- 1/2. Move place stuff to it's own directory
 
 1. Buy food
