+++ src/place/PlacePanel.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { connect } from 'react-redux'
 import ShipsForSale from './ShipsForSale'
 import Chandler from './Chandler'
-import { currentPlaceInfo } from './stateUtils'
-import { buyShip } from './actions'
+import { currentPlaceInfo } from '../stateUtils'
+import { buyShip } from '../actions'
 
 const PlacePanelComponent = ({place, cash, buyShip}) => (
     <div >
